requirements.txt
alabaster==0.7.13
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
attr==0.3.2
Babel==2.12.1
backcall==0.2.0
black==23.3.0
boltons==23.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
celery==5.3.1
comm==0.1.3
configparser==6.0.0
cryptography==41.0.1
Cython==0.29.36
debugpy==1.6.7
decorator==5.1.1
Django==4.2.3
django-extensions==3.2.3
djangorestframework==3.14.0
dl==0.1.0
docopt==0.6.2
docutils==0.20.1
exceptiongroup==1.1.2
executing==1.2.0
factory-boy==3.2.1
Faker==18.13.0
flake8==6.0.0
funcsigs==1.0.2
gunicorn==20.1.0
HTMLParser==0.0.2
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
keyring==24.2.0
lockfile==0.12.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==9.1.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipdeptree==2.9.4
pipreqs==0.4.13
platformdirs==3.8.1
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
psycopg==3.1.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
pyjnius==1.5.0
pyOpenSSL==23.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
railroad==0.5.0
redis==4.6.0
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
trove-classifiers==2023.7.6
typing_extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
widgetsnbextension==4.0.8
yarg==0.1.9
zipp==3.16.0
